fix: remove workspace and add separate folder
