--- stress-example/flows/index.yml
@@ -1,32 +1,0 @@
-!Flow
-with:
-  logserver: false
-  show_exc_info: true
-pods:
-  #crafter:
-  #  uses: pods/craft.yml
-  #  read_only: true
-  encoder:
-    uses: pods/encoder.yml
-    parallel: $JINA_PARALLEL
-    timeout_ready: 600000
-    read_only: true
-  vec_idx:
-    uses: pods/vec.yml
-    shards: $JINA_SHARDS
-    separated_workspace: true
-    polling: all
-    uses_reducing: _merge_matches
-    timeout_ready: 100000 # larger timeout as in query time will read all the data
-  doc_idx:
-    uses: pods/doc.yml
-    shards: $JINA_SHARDS
-    separated_workspace: true
-    polling: all
-    uses_reducing: _merge_matches
-    timeout_ready: 100000 # larger timeout as in query time will read all the data
-    needs: [gateway]
-  join_all:
-    uses: _merge
-    needs: [doc_idx, vec_idx]
-    read_only: true
