+++ stress-example/image_search/requirements.txt
@@ -1,0 +1,2 @@
+torch
+jina[cv,http]==0.9.12
